Abundance weighting (#129)
* improve kmer precluster help

* add abundance_weighting script

* add target_weighting script

* attempt abundance weighting

* fix weightings application

* fix tests
ensure abundance weighting changes the coassembly that would have been chosen

* total_targets as float

+++ binchicken/workflow/scripts/evaluate.py
@@ -25,7 +25,7 @@ APPRAISE_COLUMNS = SINGLEM_COLUMNS | {
 CLUSTER_COLUMNS = {
     "samples": str,
     "length": int,
-    "total_targets": int,
+    "total_targets": float,
     "total_size": float,
     "recover_samples": str,
     "coassembly": str,
